+++ CHANGELOG.md
@@ -17,6 +17,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ---
 
+## [1.0.9] - 2026-02-19
+
+### Added
+- Hello World default skill in router
+
+### Changed
+- Structure of config.json in router
+- Change start up prompt for agent and router
+
+### Fixed
+- N/A
+
+---
+
 ## [1.0.8] - 2026-02-18
 
 ### Added